serial: fix issue with 8th bit being stripped when using parity
When opening a serial port using serial_open_advanced which is
configured to be 8 data bits with a parity bit, the most significant
bit is strip from the data bits. This is caused by the ISTRIP flag
being set for the terminal attribute flags. Fix this by only setting
the ISTRIP flag when there are less than 8 data bits.

Signed-off-by: Ryan Barnett <ryan.barnett@rockwellcollins.com>
Signed-off-by: Vanya A. Sergeev <v@sergeev.io>