cmake: add cmake build support
CMake is used to control the software compilation process using simple
platform and compiler independent configuration files, and generate
native makefiles and workspaces that can be used in the compiler
environment of your choice. See: https://cmake.org/

Also with cmake support, build shared or static library, install to
$libdir, and install headers files to $includir/c-periphery.

With cmake support, it's possible to check if kernel headers is required
for build c-periphery.

Keep Makefile, because lua-periphery's build process currently expects
it.

Signed-off-by: Joris Offouga <offougajoris@gmail.com>
Signed-off-by: Vanya A. Sergeev <v@sergeev.io>