CMakeLists.txt: fix identifying COMMIT_ID within build system
When using a Buildroot (a build system for generating Embedded Linux
images) to cross compile the c-periphery package, the git commit ID
from Buildroot will be used if Buildroot itself is a Git repo. For
example, Buildroot will perform a build of c-periphery at:

  buildroot/output/build/c-cperiphery

The buildroot directory is a Git repo but the c-periphery source
folder is not. The following command:

  git describe

will use the top-level '.git' directory in buildroot for getting the
COMMIT_ID. It will then populate the buildroot COMMIT_ID for
c-periphery itself instead of putting in 'unknown'.

Utilize '--git-dir=' option to force it to only use the '.git'
directory from within the top-level c-periphery folder.

Signed-off-by: Ryan Barnett <ryan.barnett@rockwellcollins.com>