tests: add build option for tests
Add CMake build option "BUILD_TESTS" to give the ability to disable
building of the tests. The default behavior is to still build the
tests in tests/.

Create standalone CMakeLists.txt file for the test directory.
Optionally enable the building of the test_gpio.c file when the cdev
GPIO interface is available.

Signed-off-by: Ryan Barnett <ryan.barnett@rockwellcollins.com>